feat: add repl component
