+++ wi/index.html
@@ -4,7 +4,12 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>🤖 psbots</title>
+    <link rel="stylesheet" href="/src/style.css" />
     <script type="module" src="/src/main.ts"></script>
   </head>
-  <body><div id="terminal"></div></body>
+  <body>
+    <ps-bots></ps-bots>
+    <ps-bots></ps-bots>
+    <ps-bots></ps-bots>
+  </body>
 </html>
